backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
python-whois==0.8.0
dnspython==2.4.2
tldextract==5.1.1
joblib>=1.3.0
gunicorn==21.2.0
